Fix does not start unicorn

+++ config/unicorn.rb
@@ -25,7 +25,7 @@ before_fork do |server, worker|
   defined?(ActiveRecord::Base) and ActiveRecord::Base.connection.disconnect!
 
   old_pid = "#{server.config[:pid]}.oldbin"
-  unless old_pid == server.pid
+  if File.exists?(old_pid) && old_pid != server.pid
     begin
       Process.kill :QUIT, File.read(old_pid).to_i
     rescue Errno::ENONET, Errno::ESRCH
